requirements.txt
streamlit
langchain
langchain_community
langchain-huggingface
groq
sentence-transformers
faiss-cpu
pypdf
tiktoken
torch
python-dotenv
python-docx
pillow